App_Api/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
openai==0.28.1
pydantic==2.4.2
requests==2.31.0
python-dotenv==1.0.0